refactor: converge artifact proof and verification workflows
Split durable artifact inspection, observation persistence, cohort queries, artifact proof, and raw-corpus verification into explicit workflow modules while keeping the public facades stable. Introduce typed request/result models so check, QA, and reporting surfaces consume the same evidence workflow family.
