+++ polylogue/schemas/verification_requests.py
@@ -1,0 +1,40 @@
+"""Typed request objects for artifact-proof and corpus-verification workflows."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+from polylogue.protocols import ProgressCallback
+
+
+@dataclass(frozen=True)
+class ArtifactObservationQuery:
+    providers: list[str] | None = None
+    support_statuses: list[str] | None = None
+    artifact_kinds: list[str] | None = None
+    record_limit: int | None = None
+    record_offset: int = 0
+
+
+@dataclass(frozen=True)
+class ArtifactProofRequest:
+    providers: list[str] | None = None
+    record_limit: int | None = None
+    record_offset: int = 0
+
+
+@dataclass(frozen=True)
+class SchemaVerificationRequest:
+    providers: list[str] | None = None
+    max_samples: int | None = None
+    record_limit: int | None = None
+    record_offset: int = 0
+    quarantine_malformed: bool = False
+    progress_callback: ProgressCallback | None = None
+
+
+__all__ = [
+    "ArtifactObservationQuery",
+    "ArtifactProofRequest",
+    "SchemaVerificationRequest",
+]
